third_party/blink/web_tests/editing/editability/ignored-content-expected.txt

This test ensures WebKit does not ignore the content of hr, datagrid, and object with child nodes.

"world" and only "world" should be bolded in each element below:
| <hr>
|   "hello "
|   <b>
|     "world"
|   " WebKit"
| <datagrid>
|   "hello "
|   <b>
|     "world"
|   " WebKit"
| <object>
|   "hello "
|   <b>
|     "<#selection-anchor>world<#selection-focus>"
|   " WebKit"
|   <shadow:root>
|     <slot>

The result below should be identical to the result above:
| <hr>
|   "hello "
|   <b>
|     "world"
|   " WebKit"
| <datagrid>
|   "hello "
|   <b>
|     "world"
|   " WebKit"
| <object>
|   "hello "
|   <b>
|     "world"
|   " WebKit"
|   <shadow:root>
|     <slot>
